notebooks/week14_bonus/requirements.txt
click~=8.1.7
openai~=1.56.2
python-dotenv~=1.0.1
requests~=2.32.3
